Use new Framework Version format
See https://github.com/endjin/Endjin.Ip.Maturity.Matrix/issues/3 for the
changes in the IMM services that enable this. But the short version is
that this enables us to specify which particular framework is being
targetted instead of baking in a score of, say, 3. This enables the
score to drop over time automatically as new frameworks come out.